Changed @RestControllerAdvice component test logic
