--- src/test/resources/application-test.yml
@@ -1,10 +1,0 @@
-spring:
-  application:
-    name: PassManagingApplication
-  data:
-    mongodb:
-      host: localhost
-      port: 27017
-      database: test_db
-      username: root
-      password: example
